[BN-1028]: Moved test helper functions to the utils.py
